refactor(keycloak): keycloak code updation
