+++ multitenancy-rest-service/src/config/index.ts
@@ -1,3 +1,4 @@
 import app from './app';
+import keycloak from './keycloak';
 
-export default [app];
+export default [app, keycloak];
